+++ app/forms/registration_form.py
@@ -1,6 +1,7 @@
 from flask_wtf import FlaskForm
 from wtforms import StringField, SubmitField, EmailField, PasswordField
 from wtforms.validators import DataRequired, Email
+from app.utils import PasswordComplexity
 
 
 class RegistrationForm(FlaskForm):
@@ -17,5 +18,5 @@ class RegistrationForm(FlaskForm):
     first_name = StringField("Prénom", validators=[DataRequired()])
     last_name = StringField("Nom", validators=[DataRequired()])
     email = EmailField("Email", validators=[DataRequired(), Email()])
-    password = PasswordField("Mot de passe", validators=[DataRequired()])
+    password = PasswordField("Mot de passe", validators=[DataRequired(), PasswordComplexity()])
     submit = SubmitField("Demande l'inscription")
