+++ app/home/routes.py
@@ -75,10 +75,10 @@ def favorize(id_file):
     """
     try:
         if db.session.query(FICHIER).filter_by(id_Fichier=id_file).first() is None:
-                return jsonify({"error": "Ce fichier n'existe pas"}), 500
+                return jsonify({"error": "Ce fichier n'existe pas."}), 500
         if request.method == "POST":
             if not current_user.est_Actif_Utilisateur:
-                return jsonify({"error": "Vous ne pouvez pas ajouter un favori, votre compte est desactivé"}), 500
+                return jsonify({"error": "Vous ne pouvez pas ajouter un favori, votre compte est desactivé."}), 500
             db.session.execute(
                 FAVORIS.insert().values(
                     id_Fichier=id_file, id_Utilisateur=current_user.id_Utilisateur
@@ -86,7 +86,7 @@ def favorize(id_file):
             )
         else:
             if not current_user.est_Actif_Utilisateur:
-                return jsonify({"error": "Vous ne pouvez pas supprimer un favori, votre compte est desactivé"}), 500
+                return jsonify({"error": "Vous ne pouvez pas supprimer un favori, votre compte est desactivé."}), 500
             db.session.query(FAVORIS).filter(
                 FAVORIS.c.id_Fichier == id_file,
                 FAVORIS.c.id_Utilisateur == current_user.id_Utilisateur,
